moved settings to portfolio
