+++ packages/client/src/app/app.module.ts
@@ -23,10 +23,10 @@ import { IconsModule, AlertsModule, NagBarModule } from './modules';
 import { Divider, dividerHandler, TextSoftBreakBlot, shiftEnterHandler, brMatcher, textNodeMatcher } from './util/quill';
 
 import { HomeComponent, LatestComponent, WatchingPageComponent, CollectionsComponent, HistoryPageComponent,
-  BlogsComponent, InboxComponent, SettingsComponent, AlertsComponent } from './pages/home';
+  BlogsComponent, InboxComponent, AlertsComponent } from './pages/home';
 
 import { PortfolioComponent, PortHomeComponent, PortBlogComponent, PortCollectionsComponent, 
-  PortBlogPageComponent, PortCollectionPageComponent, WorksComponent } from './pages/portfolio';
+  PortBlogPageComponent, PortCollectionPageComponent, WorksComponent, SettingsComponent } from './pages/portfolio';
 
 import { BrowseComponent, GroupsComponent, NewsComponent, PostPageComponent } from './pages';
 import { RegisterComponent } from './pages/account';
