+++ packages/client/src/app/pages/home/index.ts
@@ -5,5 +5,4 @@ export { CollectionsComponent } from './collections/collections.component';
 export { HistoryComponent as HistoryPageComponent } from './history/history.component';
 export { BlogsComponent } from './blogs/blogs.component';
 export { InboxComponent } from './inbox/inbox.component';
-export { SettingsComponent } from './settings/settings.component';
 export { AlertsComponent } from './alerts/alerts.component';
