+++ packages/client/src/app/pages/portfolio/index.ts
@@ -6,3 +6,4 @@ export { PortCollectionsComponent } from './port-collections/port-collections.co
 export { PortCollectionPageComponent } from './port-collection-page/port-collection-page.component';
 
 export { WorksComponent } from './works/works.component';
+export { SettingsComponent } from './settings/settings.component';
